Unify IdentityIdpFunctions::Proof* classes with their job classes
